Random_knuth_matroid produces actually interesting matroids now

+++ src/types.jl
@@ -2,7 +2,7 @@ struct ClosedSetsMatroid{T}
   n::Integer # Size of universe
   r::Integer # Final rank (r == length(F)).
   F::Vector{Set{T}} # Closed sets by rank
-  rank::Dict{T, UInt8} # Mapping from sets to rank.
+  rank::Dict{T, Integer} # Mapping from sets to rank.
   Type::DataType
 end
 
@@ -12,7 +12,7 @@ struct FullMatroid{T}
   F::Vector{Set{T}} # Closed sets by rank
   I::Vector{Set{T}} # Independent sets by rank
   C::Set{T} # Circuits
-  rank::Dict{T, UInt8}
+  rank::Dict{T, Integer}
   Type::DataType
 end
